melanjutkan fitur pengaduan di hak akses admin

+++ app/Http/Controllers/ComplaintController.php
@@ -6,12 +6,17 @@ use App\Models\Complaint;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Validation\Rule;
 
 class ComplaintController extends Controller
 {
     public function index(){
 
-        $complaints = Complaint::where('resident_id', Auth::user()->resident->id)->paginate(5);
+
+        $residentId =  Auth::user()->resident->id ?? null;
+        $complaints = Complaint::when(Auth::user()->role_id == 2, function ($query) use($residentId){
+            $query->where('resident_id', $residentId);
+        })->paginate(5);
 
         return view('pages.complaint.index', compact(
             'complaints',
@@ -20,6 +25,11 @@ class ComplaintController extends Controller
 
     public function create()
     {
+        $resident = Auth::user()->resident;
+        if(!$resident){
+            return redirect('/complaint')->with('error', 'akun anda belum terhubung ke penduduk manapun');
+        }
+
         return view('pages.complaint.create');
     }
 
@@ -31,8 +41,14 @@ class ComplaintController extends Controller
             'photo_proof' => ['nullable', 'image', 'mimes:png,jpg,jpeg', 'max:2048'],
         ]);
 
+
+        $resident = Auth::user()->resident;
+        if(!$resident){
+            return redirect('/complaint')->with('error', 'akun anda belum terhubung ke penduduk manapun');
+        }
+
         $complaint = new Complaint();
-        $complaint->resident_id = Auth::user()->resident->id;
+        $complaint->resident_id = $resident->id;
         $complaint->title = $request->input('title');
         $complaint->content = $request->input('content');
 
@@ -48,6 +64,11 @@ class ComplaintController extends Controller
 
     public function edit($id)
     {
+        $resident = Auth::user()->resident;
+        if(!$resident){
+            return redirect('/complaint')->with('error', 'akun anda belum terhubung ke penduduk manapun');
+        }
+
         $complaint = Complaint::findOrFail($id);
 
         return view('pages.complaint.edit', compact('complaint'));
@@ -61,8 +82,19 @@ class ComplaintController extends Controller
             'photo_proof' => ['nullable', 'image', 'mimes:png,jpg,jpeg', 'max:2048'],
         ]);
 
+        $resident = Auth::user()->resident;
+        if(!$resident){
+            return redirect('/complaint')->with('error', 'akun anda belum terhubung ke penduduk manapun');
+        }
+
         $complaint = Complaint::findOrFail($id);
-        $complaint->resident_id = Auth::user()->resident->id;
+
+        if($complaint->status != 'new'){
+            return redirect('/complaint')->with('error', "Gagal mengubah aduan, status aduan anda saat ini
+            adalah $complaint->status_label");
+        }
+
+        $complaint->resident_id = $resident->id;
         $complaint->title = $request->input('title');
         $complaint->content = $request->input('content');
 
@@ -81,9 +113,37 @@ class ComplaintController extends Controller
 
     public function destroy($id)
     {
+        $resident = Auth::user()->resident;
+        if(!$resident){
+            return redirect('/complaint')->with('error', 'akun anda belum terhubung ke penduduk manapun');
+        }
+
         $complaint = Complaint::findOrFail($id);
+        if($complaint->status != 'new'){
+            return redirect('/complaint')->with('error', "Gagal menghapus aduan, status aduan anda saat ini
+            adalah $complaint->status_label");
+        }
+
         $complaint->delete();
 
         return redirect('/complaint')->with('success', 'Berhasil menghapus aduan');
+    }
+
+    public function update_status(Request $request, $id)
+    {
+        $request->validate([
+            'status' => ['required', Rule::in(['new', 'processing', 'completed'])],
+        ]);
+
+        $resident = Auth::user()->resident;
+        if(Auth::user()->role_id == 2 && !$resident){
+            return redirect('/complaint')->with('error', 'akun anda belum terhubung ke penduduk manapun');
+        }
+
+        $complaint = Complaint::findOrFail($id);
+        $complaint->status = $request->input('status');
+        $complaint->save();
+
+        return redirect('/complaint')->with('success', 'Berhasil mengubah status');
     }
 }
